# docs/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MADAVOLA API
  version: v1
  description: API contract v1 (spec-first). Certains endpoints sont planifiés.
servers:
  - url: /api/v1
...
paths:
...
  /transactions:
    get:
      tags: [transactions]
      summary: List transactions
      parameters:
        - in: query
          name: seller_actor_id
          schema: { type: integer }
        - in: query
          name: buyer_actor_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: #/components/schemas/Transaction
    post:
      tags: [transactions]
      summary: Create transaction
      x-implementation-status: planned
      responses:
        "201":
          description: Created
  /transactions/{transaction_id}:
    get:
      tags: [transactions]
      summary: Get transaction
      parameters:
        - in: path
          name: transaction_id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: #/components/schemas/TransactionDetail
  /transactions/{transaction_id}/payments:
    get:
      tags: [transactions]
      summary: List payments for transaction
      parameters:
        - in: path
          name: transaction_id
          required: true
          schema: { type: integer }
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: #/components/schemas/TransactionPayment
  /transactions/{transaction_id}/initiate-payment:
    post:
      tags: [transactions]
      summary: Initiate payment for transaction
      parameters:
        - in: path
          name: transaction_id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: #/components/schemas/TransactionPaymentInitiate
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: #/components/schemas/TransactionPayment
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id: { type: integer }
        seller_actor_id: { type: integer }
        buyer_actor_id: { type: integer }
        status: { type: string }
        total_amount: { type: number }
        currency: { type: string }
...
    TransactionPaymentInitiate:
      type: object
      properties:
        provider_code: { type: string }
        external_ref: { type: string }
        idempotency_key: { type: string }
      required: [provider_code]
    TransactionPayment:
      type: object
      properties:
        payment_request_id: { type: integer }
        payment_id: { type: integer }
        status: { type: string }
        external_ref: { type: string }
...
    TransactionItem:
      type: object
      properties:
        id: { type: integer }
        lot_id: { type: integer, nullable: true }
        quantity: { type: number }
        unit_price: { type: number }
        line_amount: { type: number }
    TransactionDetail:
      type: object
      properties:
        id: { type: integer }
        seller_actor_id: { type: integer }
        buyer_actor_id: { type: integer }
        status: { type: string }
        total_amount: { type: number }
        currency: { type: string }
        items:
          type: array
          items:
            $ref: #/components/schemas/TransactionItem
...